refactor cli entry
